# Optimizations based on the array being sorted