I tried telling the story of what I did through the commits.

My times:                 | Safari (13.0.2) | Chrome (77.0.3865.120) | Node (v12.12.0)
:------------------------ | :-------------- | :--------------------- | :--------------
useFilter                 | 246 ms          | 144 ms                 | 146 ms
useSet                    | 1071 ms         | 120 ms                 | 117 ms
useReduce                 | 284 ms          | 215 ms                 | 198 ms
useForIncludes            | 465 ms          | 404 ms                 | 288 ms
useAssociative            | 652 ms          | 263 ms                 | 257 ms
justCopy                  | 94 ms           | 33 ms                  | 38 ms
copyAndSort               | 1917 ms         | 1402 ms                | 1254 ms
useIterator               | 109 ms          | 23 ms                  | 20 ms
useIteratorWisely         | 86 ms           | 19 ms                  | 20 ms
useBetterReduce           | 135 ms          | 141 ms                 | 154 ms
useBetterReduce_backwards | 128 ms          | 162 ms                 | 143 ms